Separate reboot tests
